Routes re-structured, navigation/footer moved to shared components.
