+++ src/ignitus-Routes/publicRoutes/index.js
@@ -1,26 +1,26 @@
 import React from "react";
 import { Switch, Route } from "react-router-dom";
-import Navigation from "../ignitus-Navigation";
-import Footer from "../ignitus-Footer";
+import Navigation from "../../ignitus-Shared/Components/ignitus-Navigation";
+import Footer from "../../ignitus-Shared/Components/ignitus-Footer";
 
-import Contributors from "../ignitus-HomePage/ignitus-Contributors";
-import GetStarted from "../ignitus-HomePage/ignitus-GetStarted";
-import Introduction from "../ignitus-HomePage/ignitus-Introduction";
-import Testimonial from "../ignitus-HomePage/ignitus-Testimonial";
-import WhatWeProvide from "../ignitus-HomePage/ignitus-WhatWeDo";
-import Partner from "../ignitus-HomePage/ignitus-Partners";
+import Contributors from "../../ignitus-HomePage/ignitus-Contributors";
+import GetStarted from "../../ignitus-HomePage/ignitus-GetStarted";
+import Introduction from "../../ignitus-HomePage/ignitus-Introduction";
+import Testimonial from "../../ignitus-HomePage/ignitus-Testimonial";
+import WhatWeProvide from "../../ignitus-HomePage/ignitus-WhatWeDo";
+import Partner from "../../ignitus-HomePage/ignitus-Partners";
 
-import LoginProfessor from "../ignitus-Authentication/ignitus-ProfessorLogin";
-import SignupProfessor from "../ignitus-Authentication/ignitus-ProfessorSignUp";
-import LoginStudent from "../ignitus-Authentication/ignitus-StudentLogin";
-import SignupStudent from "../ignitus-Authentication/ignitus-StudentSignUp";
-import { CommonLoginUI, CommonSignUpUI } from "../ignitus-Authentication/ignitus-Common-Auth-Student-Prof/Components/commonLoginRegister";
+import LoginProfessor from "../../ignitus-Authentication/ignitus-ProfessorLogin";
+import SignupProfessor from "../../ignitus-Authentication/ignitus-ProfessorSignUp";
+import LoginStudent from "../../ignitus-Authentication/ignitus-StudentLogin";
+import SignupStudent from "../../ignitus-Authentication/ignitus-StudentSignUp";
+import { CommonLoginUI, CommonSignUpUI } from "../../ignitus-Authentication/ignitus-Common-Auth-Student-Prof/Components/commonLoginRegister";
 
-import loader from "../ignitus-Assets/Images/loader2.gif";
+import loader from "../../ignitus-Assets/Images/loader2.gif";
 import Loadable from "react-loadable";
 
-import {Notfound} from './notFound';
-import JobConfirmation  from "../ignitus-Careers/ignitus-JobConfirmation";
+import { Notfound } from '../../ignitus-Shared/Components/NotFoundComponent/notFound';
+import JobConfirmation  from "../../ignitus-Careers/ignitus-JobConfirmation";
 
 // RouteBased CodeSplitting.
 
@@ -33,7 +33,7 @@ const Loading = () => (
 );
 
 const AboutUs = Loadable({
-  loader: () => import("../ignitus-AboutPage/ignitus-About/Components/About"),
+  loader: () => import("../../ignitus-AboutPage/ignitus-About/Components/About"),
   loading: Loading
 });
