+++ src/ignitus-Shared/Components/NotFoundComponent/Styles/style.scss
