+++ src/ignitus-Shared/Components/NotFoundComponent/notFound.js
@@ -1,4 +1,3 @@
-
 import React from 'react';
 import './Styles/style.scss';
 import { Link } from 'react-router-dom';
@@ -9,11 +8,11 @@ export const Notfound = () => (
       <p>404</p>
       <p>We could not find what you were looking for.</p>
       <p>
-		Please check your URL or return&nbsp;
+        Please check your URL or return&nbsp;
         <Link to="/">
           <span className="_link">home</span>
         </Link>
-					.
+        .
       </p>
     </div>
   </div>
