+++ src/ignitus-Shared/Components/ignitus-Footer/Styles/style.scss
